refactor: remover messageHelper legado e substituir por nova implementação
- Removido o arquivo antigo messageHelper.js.
- Adicionado o novo arquivo mediaHelper.js para processamento e download de mídias.
- Criado um novo arquivo messageHelper.js com melhorias no processamento de mensagens e comandos.
- Aprimorado o tratamento de erros e o sistema de logs nas novas implementações.
- Atualizada a estrutura do projeto para separar melhor as responsabilidades entre processamento de mídias e mensagens.
